put logic back to EarnTri
